feat: fix of naming and errors found by Dart Analyzer.
