--- 354313932/lib/pages/home/components/Body.dart
@@ -1,97 +1,0 @@
-import 'package:chat_journal/constants.dart';
-import 'package:chat_journal/models/Page.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-
-class Body extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    Size size = MediaQuery.of(context).size;
-    return SingleChildScrollView(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: <Widget> [
-          Opacity(
-            opacity: .4,
-            child: Container(
-              margin: EdgeInsets.symmetric(horizontal: kDefaultPadding * 3),
-              height: 50.0,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.all(Radius.circular(16.0)),
-                color: Colors.white,
-              ),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: <Widget>[
-                    Container(
-                      height: 24,
-                      width: 24,
-                      child: SvgPicture.asset("assets/icons/robot.svg"),
-                    ),
-                  Padding(
-                    padding: const EdgeInsets.only(left: 8.0),
-                    child: Text(
-                        'Questionnaire Bot',
-                        style: TextStyle(
-                          fontSize: 18.0,
-                          fontWeight: FontWeight.bold,
-                          color: Colors.black,
-                        ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          SizedBox(
-            height: size.height,
-            child: Stack(
-              children: <Widget>[
-                Container(
-                  margin: EdgeInsets.only(top: 16.0),
-                  padding: EdgeInsets.only(
-                    top: size.height * 0.12,
-                    left: kDefaultPadding,
-                    right: kDefaultPadding,
-                  ),
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.only(
-                      topLeft: Radius.circular(16),
-                      topRight: Radius.circular(16),
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: const EdgeInsets.only(top: 16.0),
-                  child: ListView.builder(
-                      itemCount: pages.length,
-                      itemBuilder: (context, index) {
-                        return ListTile(
-                          leading: Container(
-                            height: 50,
-                            width: 50,
-                            decoration: BoxDecoration(
-                              shape: BoxShape.circle,
-                              color: Color(0xFF1E81F5),
-                            ),
-                            child: pages[index].icon,
-                          ),
-                          title: Text(pages[index].title),
-                          subtitle: Text(pages[index].subtitle),
-                          onTap: () {
-                            print("ListTile $index tapped");
-                          },
-                        );
-                      },
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
